requirements.txt
absl-py==0.6.1
Keras==2.2.4
tensorflow==1.12.0
numpy==1.15.4
sklearn==0.0
scikit-image==0.14.1
tqdm==4.28.1
scipy==1.2.0
tqdm==4.28.1